REST/requirements.txt
uvicorn[standard]
fastapi==0.75.1
grpcio
grpcio-tools==1.44.0
python-multipart==0.0.5
numpy==1.22.3
locust==2.8.6